+++ CHANGES.rst
@@ -19,6 +19,9 @@ Changelog
 - Managing missing values for mail_type and assigned_user on IImioDmsIncomingMail
   [sgeulette]
 
+- Added batch actions on task dashboard
+  [sgeulette]
+
 - Improvements: disable own delete on contacts, block parent portlets on contacts, add local roles for dir_general
                 on contacts, corrected disabled treating_groups bug, updated voc cache, corrected transition batch
                 action,